argschema overhaul
